Sync latest hafs.v2.0.3 changes from NCO/SPA (Wei Wei), including * HAFSv2 related ecf script updates * Update versions/run.ver * Add dbnalert for NHC requested subsetted grib2 files * Update ush/hafs_lnchk.sh to check/warn symbolic pointing to an empty file

+++ ush/hafs_lnchk.sh
@@ -27,7 +27,11 @@ else
    symlink=$3
 fi
 
-ln -sf $2 ${symlink}
+if [ "$1" = "-r" -o -s $2 ] ; then
+  ln -sf $2 ${symlink}
+else
+  echo "WARNING: $2 does not exist or is empty."
+fi
 
 if [ "$1" = "-r" ]; then
   # No checks, same as ln -sf
@@ -46,7 +50,7 @@ elif [ "$1" = "-d" ]; then
 elif [ "$1" = "-e" ]; then
   if [ ! -e ${symlink} ]; then
     echo "FATAL ERROR: Broken symbolic link: ${symlink} => $2. Exiting."
-	err_exit "$2 does not exist. Exiting."
+	err_exit "$2 does not exist or is empty. Exiting."
   fi
 else
   echo "FATAL ERROR: Unknown command line option: $1."
